requirements.txt
hamolicious-vector-class==1.2.0
pygame==2.1.0
